feat: files [WIP]
Allow defining local files as resources:

```rust
bring fs;
new fs.JsonFile("hi.json", obj: {});
```

### `wingsdk`

* Move `Resource` from `cloud` to `core` (because it is not a cloud thing).
* Change the base class of `FileBase` to `Resource` (instead of `Construct`).
* Move `FileBase` to `core` (next to `Files`) to avoid cyclic deps between `core` and `fs`.

### `wingc`

Add `fs` and `core` to hack in `add_module_to_env` so we can do `bring fs`.

---

### Current status

I am getting the following error:

```
thread 'main' panicked at 'Encountered JSII type @winglang/wingsdk.core.FileBase which isn't part of imported JSII namespace @winglang/wingsdk.fs', libs/wingc/src/type_check/jsii_importer.rs:103:13
```
