+++ examples/tests/valid/files.w
@@ -1,0 +1,3 @@
+bring fs;
+
+new fs.JsonFile("hi.json", obj: {});
