+++ libs/wingsdk/src/cloud/counter.ts
@@ -1,7 +1,6 @@
 import { Construct, IConstruct } from "constructs";
 import { Polycons } from "polycons";
-import { CaptureMetadata, Code } from "../core";
-import { Resource } from "./resource";
+import { CaptureMetadata, Code, Resource } from "../core";
 
 /**
  * Global identifier for `Counter`.
