+++ libs/wingsdk/src/cloud/index.ts
@@ -3,4 +3,3 @@ export * from "./counter";
 export * from "./function";
 export * from "./logger";
 export * from "./queue";
-export * from "./resource";
