+++ libs/wingsdk/src/core/files.ts
@@ -1,14 +1,68 @@
 import {
   existsSync,
+  mkdirSync,
   readFileSync,
   rmSync,
   unlinkSync,
   writeFileSync,
 } from "fs";
-import { isAbsolute, join } from "path";
-import { IConstruct } from "constructs";
-import { FileBase } from "../fs";
+import { isAbsolute, dirname, join } from "path";
+import { Construct, IConstruct } from "constructs";
 import { IApp } from "./app";
+import { CaptureMetadata, Code, NodeJsCode } from "./inflight";
+import { Resource } from "./resource";
+
+/**
+ * Represents a file to be synthesized in the app's output directory.
+ */
+export abstract class FileBase extends Resource {
+  /**
+   * The file's relative path to the output directory.
+   */
+  public readonly filePath: string;
+
+  /**
+   * Indicates that generated files are not stateful. They can always be
+   * regenerated from their definition.
+   */
+  public readonly stateful = false;
+
+  /**
+   * Defines a file
+   * @param scope construct scope
+   * @param id construct id
+   * @param filePath relative file path
+   * @param props initialization props
+   */
+  constructor(scope: Construct, id: string, filePath: string) {
+    super(scope, id);
+    this.filePath = filePath;
+  }
+
+  /**
+   * Render the contents of the file and save it to the user's file system.
+   */
+  public save(outdir: string) {
+    const data = this.render();
+    const outpath = join(outdir, this.filePath);
+    mkdirSync(dirname(outpath), { recursive: true });
+    writeFileSync(outpath, data);
+  }
+
+  /**
+   * Returns the contents of the file to save.
+   */
+  protected abstract render(): string;
+
+  /**
+   * @internal
+   */
+  public _bind(captureScope: IConstruct, metadata: CaptureMetadata): Code {
+    captureScope;
+    metadata;
+    return NodeJsCode.fromInline(JSON.stringify(this.render()));
+  }
+}
 
 /**
  * Props for `Files`.
@@ -21,7 +75,7 @@ export interface FilesProps {
 
   /**
    * The path to a state file which will track all synthesized files. If a
-   * statefile is not specified, we won't be able to remove extrenous files.
+   * statefile is not specified, we won't be able to remove extraneous files.
    * @default - no state file
    */
   readonly stateFile?: string;
