+++ libs/wingsdk/src/core/index.ts
@@ -3,4 +3,5 @@ export * from "./dependency";
 export * from "./duration";
 export * from "./files";
 export * from "./inflight";
+export * from "./resource";
 export * from "./testing";
