+++ libs/wingsdk/src/fs/index.ts
@@ -1,3 +1,2 @@
-export * from "./file-base";
 export * from "./json-file";
 export * from "./text-file";
