+++ proj.w
@@ -1,0 +1,3 @@
+bring fs;
+
+new fs.JsonFile("test.json");
